fix: functions that call Func_33f2
Func_33f2 starts the event script loop, which treats the next chunk of bytes as data. The functions fixed in this commit are new to this PR and all either (a) generated fake ASM out of this data or (b) hung forever and would not disassemble.

also was able to fix an erroneous change to decompress.asm with this